Change the initial bandwidth calculation to type 'double'
It needs to be stored in 'double' format and then rounded off to
the nearest integer and returned.

+++ lib/lstbt_v.c
@@ -705,9 +705,9 @@ static void dump_adapters_num(const char *router)
 
 static u32 usb3_bw_to_mbps(u32 bw, u16 scale)
 {
-	u32 mbps;
+	double mbps;
 
-	mbps = (((bw * 512) << scale) * 8000) / (1000 * 1000);
+	mbps = ((double)((bw * 512) << scale) * 8000) / (1000 * 1000);
 	return round(mbps);
 }
